renamed validator for id, placed db file on the root level in the project
