+++ src/repositories/blogs-db-repository.ts
@@ -6,7 +6,7 @@ import {
 import { db } from "../temporal-database/project-db";
 import { transformBlogsResponse } from "../utils/blogs-utils/transformBlogsResponse";
 import { creationDate } from "../utils/common-utils/creation-publication-dates";
-import { blogsCollection } from "./db";
+import { blogsCollection } from "../db";
 import { ObjectId } from "mongodb";
 
 export const blogsList = db.blogs;
@@ -17,13 +17,13 @@ export const blogsRepository = {
     );
   },
   async findBlogById(id: string): Promise<BlogViewModel | null> {
-    const foundBlog = await blogsCollection.findOne<BlogDBType>({
-      _id: new ObjectId(id),
-    });
-    if (foundBlog) {
-      return transformBlogsResponse(foundBlog);
-    }
-    return foundBlog;
+      const foundBlog = await blogsCollection.findOne<BlogDBType>({
+        _id: new ObjectId(id),
+      });
+      if (foundBlog) {
+        return transformBlogsResponse(foundBlog);
+      }
+      return foundBlog;
   },
   async createNewBlog(body: BlogInputModel): Promise<BlogViewModel> {
     const { name, description, websiteUrl } = body;
